adv delete

+++ app/src/main/java/wxm/KeepAccount/ui/utility/AdapterNoteDetail.java
@@ -1,15 +1,14 @@
 package wxm.KeepAccount.ui.utility;
 
-import android.annotation.SuppressLint;
+import android.app.Dialog;
 import android.content.Context;
 import android.content.Intent;
+import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.RelativeLayout;
 import android.widget.SimpleAdapter;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -20,6 +19,7 @@ import wxm.KeepAccount.ui.base.FrgUitlity.FrgAdvBase;
 import wxm.KeepAccount.ui.base.Helper.ViewHelper;
 import wxm.KeepAccount.ui.base.SwipeLayout.SwipeLayout;
 import wxm.KeepAccount.utility.ContextUtil;
+import wxm.KeepAccount.utility.ToolUtil;
 import wxm.androidutil.util.FastViewHolder;
 import wxm.androidutil.util.UtilFun;
 import wxm.KeepAccount.R;
@@ -28,7 +28,6 @@ import wxm.KeepAccount.define.GlobalDef;
 import wxm.KeepAccount.define.INote;
 import wxm.KeepAccount.define.IncomeNoteItem;
 import wxm.KeepAccount.define.PayNoteItem;
-import wxm.KeepAccount.ui.base.Helper.ResourceHelper;
 import wxm.KeepAccount.ui.data.edit.Note.ACPreveiwAndEdit;
 
 /**
@@ -69,45 +68,61 @@ public class AdapterNoteDetail extends SimpleAdapter {
     public View getView(int position, View convertView, ViewGroup parent) {
         FastViewHolder vh = FastViewHolder.get(mCTSelf, convertView, R.layout.li_data_detail_holder);
         HashMap<String, INote> hmData = UtilFun.cast_t(getItem(position));
+        INote itemData = hmData.get(K_NODE);
 
         SwipeLayout sl = vh.getView(R.id.swipe);
         vwTag vt = (vwTag)sl.getTag();
         if(null == vt) {
             vt = new vwTag();
-            vt.mContent = LayoutInflater.from(mCTSelf).inflate(R.layout.li_data_detail, null);
-            vt.mRight = LayoutInflater.from(mCTSelf).inflate(R.layout.vw_delte, null);
-            vt.mRight.setTag(hmData.get(K_NODE));
-            vt.mRight.setOnClickListener(v -> {
-                INote data = (INote)v.getTag();
-                List<Integer> al = Collections.singletonList(data.getId());
+            if(itemData.isPayNote()) {
+                vt.mContent = LayoutInflater.from(mCTSelf).inflate(R.layout.liit_data_pay, null);
+                initPay(vt.mContent, itemData);
+            } else  {
+                vt.mContent = LayoutInflater.from(mCTSelf).inflate(R.layout.liit_data_income, null);
+                initIncome(vt.mContent, itemData);
+            }
 
-                if(data.isPayNote())    {
-                    ContextUtil.getPayIncomeUtility().deletePayNotes(al);
-                } else  {
-                    ContextUtil.getPayIncomeUtility().deleteIncomeNotes(al);
-                }
+            vt.mRight = LayoutInflater.from(mCTSelf).inflate(R.layout.liit_data_swipe, null);
+            ViewHelper vhRight = new ViewHelper(vt.mRight);
+            vhRight.setTag(R.id.iv_delete, itemData);
+            vhRight.setTag(R.id.iv_edit, itemData);
+
+            vhRight.getChildView(R.id.iv_delete).setOnClickListener(v -> {
+                Dialog alertDialog = new AlertDialog.Builder(mCTSelf).
+                        setTitle("删除数据").
+                        setMessage("此操作不能恢复，是否继续操作!").
+                        setPositiveButton("是", (dialog, which) -> {
+                            INote data = (INote)v.getTag();
+                            List<Integer> al = Collections.singletonList(data.getId());
+
+                            if(data.isPayNote())    {
+                                ContextUtil.getPayIncomeUtility().deletePayNotes(al);
+                            } else  {
+                                ContextUtil.getPayIncomeUtility().deleteIncomeNotes(al);
+                            }
+                        }).
+                        setNegativeButton("否", (dialog, which) -> {}).
+                        create();
+                alertDialog.show();
+            });
+
+            vhRight.getChildView(R.id.iv_edit).setOnClickListener(v -> {
+                Intent intent;
+                INote data = (INote)v.getTag();
+                intent = new Intent(mCTSelf, ACPreveiwAndEdit.class);
+                intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_ID, data.getId());
+                intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_TYPE,
+                        data.isPayNote() ? GlobalDef.STR_RECORD_PAY : GlobalDef.STR_RECORD_INCOME);
+
+                mCTSelf.startActivity(intent);
             });
 
             sl.setContentView(vt.mContent);
             sl.setRightView(vt.mRight);
-
             sl.setTag(vt);
         }
 
-        initInfo(vt.mContent, hmData);
         return vh.getConvertView();
-    }
-
-
-    private void initInfo(View vwInfo, HashMap<String, INote> hmData)  {
-        INote data = hmData.get(K_NODE);
-        if (data.isPayNote()) {
-            vwInfo.findViewById(R.id.rl_income).setVisibility(View.GONE);
-            initPay(vwInfo, data);
-        } else {
-            vwInfo.findViewById(R.id.rl_pay).setVisibility(View.GONE);
-            initIncome(vwInfo, data);
-        }
     }
 
     /**
@@ -135,15 +150,6 @@ public class AdapterNoteDetail extends SimpleAdapter {
 
         // for look detail
         vHelper.setVisibility(R.id.iv_pay_action, View.GONE);
-
-        vHelper.getChildView(R.id.rl_pay).setOnClickListener(v -> {
-            Intent intent;
-            intent = new Intent(mCTSelf, ACPreveiwAndEdit.class);
-            intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_ID, data.getId());
-            intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_TYPE, GlobalDef.STR_RECORD_PAY);
-
-            mCTSelf.startActivity(intent);
-        });
 
         // for budget
         if (UtilFun.StringIsNullOrEmpty(b_name)) {
@@ -175,15 +181,6 @@ public class AdapterNoteDetail extends SimpleAdapter {
 
         // for look detail
         vHelper.setVisibility(R.id.iv_income_action, View.GONE);
-
-        vHelper.getChildView(R.id.rl_income).setOnClickListener(v -> {
-            Intent intent;
-            intent = new Intent(mCTSelf, ACPreveiwAndEdit.class);
-            intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_ID, data.getId());
-            intent.putExtra(GlobalDef.INTENT_LOAD_RECORD_TYPE, GlobalDef.STR_RECORD_INCOME);
-
-            mCTSelf.startActivity(intent);
-        });
 
         // for note
         String nt = i_n.getNote();
